dns-spoof-detector/requirements.txt
dnspython
ipwhois
requests
flask